+++ client/src/components/AppBar/dropDownBackup.js
@@ -1,0 +1,60 @@
+import React, { Component } from 'react';
+import Button from '@material-ui/core/Button';
+import Menu from '@material-ui/core/Menu';
+import MenuItem from '@material-ui/core/MenuItem';
+import ArrowDropDownIcon from '@material-ui/icons/ArrowDropDown';
+
+export default class Dropdown extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            anchorEl: null,
+            currentViewName: this.props.currentViewName,
+        };
+        this.handleClick = this.handleClick.bind(this);
+        this.handleClose = this.handleClose.bind(this);
+    }
+
+    handleClick = (event) => {
+        this.setState({
+            anchorEl: event.currentTarget,
+        });
+    };
+
+    handleClose = (event) => {
+        this.setState({
+            anchorEl: null,
+        });
+        if (event.currentTarget.title) {
+            this.props.currentViewNameChange(event.currentTarget.title);
+        }
+    };
+
+    render() {
+        return (
+            <div>
+                <Button
+                    aria-controls="simple-menu"
+                    aria-haspopup="true"
+                    onClick={this.handleClick}
+                    variant="outlined"
+                    size="small"
+                    endIcon={<ArrowDropDownIcon />}
+                >
+                    {this.state.currentViewName}
+                </Button >
+                <Menu
+                    id="simple-menu"
+                    anchorEl={this.state.anchorEl}
+                    keepMounted
+                    open={Boolean(this.state.anchorEl)}
+                    onClose={this.handleClose}
+                >
+                    <MenuItem title="Day" onClick={this.handleClose}>Day</MenuItem>
+                    <MenuItem title="Week" onClick={this.handleClose}>Week</MenuItem>
+                    <MenuItem title="Month" onClick={this.handleClose}>Month</MenuItem>
+                </Menu>
+            </div>
+        )
+    }
+}
