+++ client/src/components/pickerBackup.js
@@ -1,44 +1,54 @@
 import format from 'date-fns/format';
-import React ,{useState}from 'react';
+import React, { Component } from 'react';
 import Button from '@material-ui/core/Button';
 import Hidden from '@material-ui/core/Hidden';
 import ArrowDropDownIcon from '@material-ui/icons/ArrowDropDown';
 import DateFnsUtils from '@date-io/date-fns';
 import { MuiPickersUtilsProvider, DatePicker, } from '@material-ui/pickers';
-import {useSelector,useDispatch} from 'react-redux'
-import {changeCurrentDate} from '../redux/actions/index'
 
-const Picker = () => {
-    const dispatch    = useDispatch()
-    const currentDate = useSelector(state => state.currentDate.date)
-    const [pickerIsOpen, setPickerIsOpen] = useState(false)
+export default class Picker extends Component {
 
-    const handleSelcetedDate = () => {
-        dispatch(changeCurrentDate(currentDate))
+    constructor(props) {
+        super(props);
+        this.state = {
+            currentDate: this.props.currentDate,
+            currentViewName: this.props.currentViewName,
+            pickerIsOpen: false,
+        };
     }
 
-
-
-    let pickerFormat = "MMMM yyyy"
-    if(currentDate === "Day"){
-        pickerFormat = "d MMMM yyyy"
+    setPicker = (pickerIsOpen) => {
+        this.setState({ pickerIsOpen });
     }
-    let date = format(currentDate,pickerFormat)
-    
-    return (
-        <div>
+
+    handleDateChange = (currentDate) => {
+        this.setState({ currentDate });
+        this.props.handleSelectedDate(this.state.currentDate);
+    }
+
+    render() {
+
+        let pickerFormat = "MMMM yyyy";
+        let date = format(this.state.currentDate, 'MMM yyyy');
+        if (this.state.currentViewName === "Day") {
+            pickerFormat = "d MMMM yyyy";
+            date = format(this.state.currentDate, 'd MMM yyyy');
+        }
+
+        return (
+            <div>
                 {
-                    pickerIsOpen
+                    this.state.pickerIsOpen
                         ? <MuiPickersUtilsProvider utils={DateFnsUtils}>
                             <DatePicker
                                 variant="dialog"
                                 format={pickerFormat}
                                 disableToolbar={false}
-                                value={currentDate}
-                                onChange={handleSelcetedDate}
-                                open={pickerIsOpen}
-                                onOpen={() => { setPickerIsOpen(true) }}
-                                onClose={() => { setPickerIsOpen(false) }}
+                                value={this.state.currentDate}
+                                onChange={this.props.handleSelectedDate}
+                                open={this.state.pickerIsOpen}
+                                onOpen={() => { this.setPicker(true) }}
+                                onClose={() => { this.setPicker(false) }}
                                 KeyboardButtonProps={{ 'aria-label': 'change date', }}
                             />
                         </MuiPickersUtilsProvider>
@@ -46,7 +56,7 @@ const Picker = () => {
                             <Hidden smUp>
                                 <Button
                                     endIcon={<ArrowDropDownIcon />}
-                                    onClick={() => { setPickerIsOpen(true) }}
+                                    onClick={() => { this.setPicker(true) }}
                                     style={{ color: "#616161" }}
                                 >
                                     {date}
@@ -54,7 +64,7 @@ const Picker = () => {
                             </Hidden>
                             <Hidden xsDown>
                                 <Button
-                                    onClick={() => { setPickerIsOpen(true) }}
+                                    onClick={() => { this.setPicker(true) }}
                                     style={{ color: "#616161" }}
                                 >
                                     {date}
@@ -65,7 +75,6 @@ const Picker = () => {
 
                 }
             </div>
-    )
+        )
+    }
 }
-
-export default Picker
